Move tuple handling out of DynamicDataAttribute
