+++ src/TestFramework/TestFramework/Attributes/DataSource/DynamicDataOperations.cs
@@ -124,10 +124,7 @@ internal static class DynamicDataOperations
                     return false;
                 }
 
-                if (!TryHandleTupleDataSource(entry, objects))
-                {
-                    objects.Add(new[] { entry });
-                }
+                objects.Add(new[] { entry });
             }
 
             data = objects;
@@ -137,163 +134,6 @@ internal static class DynamicDataOperations
         data = null;
         return false;
     }
-
-    private static bool TryHandleTupleDataSource(object data, List<object[]> objects)
-    {
-#if NET471_OR_GREATER || NETCOREAPP
-        if (data is ITuple tuple
-            && (objects.Count == 0 || objects[^1].Length == tuple.Length))
-        {
-            object[] array = new object[tuple.Length];
-            for (int i = 0; i < tuple.Length; i++)
-            {
-                array[i] = tuple[i]!;
-            }
-
-            objects.Add(array);
-            return true;
-        }
-#else
-        if (IsTupleOrValueTuple(data, out int tupleSize)
-            && (objects.Count == 0 || objects[objects.Count - 1].Length == tupleSize))
-        {
-            object[] array = new object[tupleSize];
-            ProcessTuple(data, array, 0);
-
-            objects.Add(array);
-            return true;
-        }
-
-        static object GetFieldOrProperty(Type type, object data, string fieldOrPropertyName)
-            // ValueTuple is a value type, and uses fields for Items.
-            // Tuple is a reference type, and uses properties for Items.
-            => type.IsValueType
-                ? type.GetField(fieldOrPropertyName).GetValue(data)
-                : type.GetProperty(fieldOrPropertyName).GetValue(data);
-
-        static void ProcessTuple(object data, object[] array, int startingIndex)
-        {
-            Type type = data.GetType();
-            int tupleSize = type.GenericTypeArguments.Length;
-            for (int i = 0; i < tupleSize; i++)
-            {
-                if (i != 7)
-                {
-                    // Note: ItemN are properties on Tuple, but are fields on ValueTuple
-                    array[startingIndex + i] = GetFieldOrProperty(type, data, $"Item{i + 1}");
-                    continue;
-                }
-
-                object rest = GetFieldOrProperty(type, data, "Rest");
-                if (IsTupleOrValueTuple(rest, out _))
-                {
-                    ProcessTuple(rest, array, startingIndex + 7);
-                }
-                else
-                {
-                    array[startingIndex + i] = rest;
-                }
-
-                return;
-            }
-        }
-#endif
-
-        return false;
-    }
-
-#if !NET471_OR_GREATER && !NETCOREAPP
-    private static bool IsTupleOrValueTuple(object? data, out int tupleSize)
-    {
-        tupleSize = 0;
-
-        if (data is null)
-        {
-            return false;
-        }
-
-        Type type = data.GetType();
-        if (!type.IsGenericType)
-        {
-            return false;
-        }
-
-        Type genericTypeDefinition = type.GetGenericTypeDefinition();
-
-        if (genericTypeDefinition == typeof(Tuple<>) ||
-            genericTypeDefinition == typeof(Tuple<,>) ||
-            genericTypeDefinition == typeof(Tuple<,,>) ||
-            genericTypeDefinition == typeof(Tuple<,,,>) ||
-            genericTypeDefinition == typeof(Tuple<,,,,>) ||
-            genericTypeDefinition == typeof(Tuple<,,,,,>) ||
-            genericTypeDefinition == typeof(Tuple<,,,,,,>))
-        {
-            tupleSize = type.GetGenericArguments().Length;
-            return true;
-        }
-
-        if (genericTypeDefinition == typeof(Tuple<,,,,,,,>))
-        {
-            object? last = type.GetProperty("Rest").GetValue(data);
-            if (IsTupleOrValueTuple(last, out int restSize))
-            {
-                tupleSize = 7 + restSize;
-                return true;
-            }
-            else
-            {
-                tupleSize = 8;
-                return true;
-            }
-        }
-
-#if NET462
-        if (genericTypeDefinition.FullName.StartsWith("System.ValueTuple`", StringComparison.Ordinal))
-        {
-            tupleSize = type.GetGenericArguments().Length;
-            if (tupleSize == 8)
-            {
-                object? last = type.GetField("Rest").GetValue(data);
-                if (IsTupleOrValueTuple(last, out int restSize))
-                {
-                    tupleSize = 7 + restSize;
-                }
-            }
-
-            return true;
-        }
-#else
-        if (genericTypeDefinition == typeof(ValueTuple<>) ||
-            genericTypeDefinition == typeof(ValueTuple<,>) ||
-            genericTypeDefinition == typeof(ValueTuple<,,>) ||
-            genericTypeDefinition == typeof(ValueTuple<,,,>) ||
-            genericTypeDefinition == typeof(ValueTuple<,,,,>) ||
-            genericTypeDefinition == typeof(ValueTuple<,,,,,>) ||
-            genericTypeDefinition == typeof(ValueTuple<,,,,,,>))
-        {
-            tupleSize = type.GetGenericArguments().Length;
-            return true;
-        }
-
-        if (genericTypeDefinition == typeof(ValueTuple<,,,,,,,>))
-        {
-            object? last = type.GetField("Rest").GetValue(data);
-            if (IsTupleOrValueTuple(last, out int restSize))
-            {
-                tupleSize = 7 + restSize;
-                return true;
-            }
-            else
-            {
-                tupleSize = 8;
-                return true;
-            }
-        }
-#endif
-
-        return false;
-    }
-#endif
 
     private static PropertyInfo? GetPropertyConsideringInheritance(Type type, string propertyName)
     {
